Restore babel.js
